(bot) losses fully work

+++ src/analytics/handlers/layout_util.py
@@ -13,7 +13,8 @@ async def enter_step(msg_data: MsgData, step: int, branch: str) -> None:
     state_data = await msg_data.state.get_data()
     
     messages_to_delete = state_data.get("report:messages_to_delete")
-    msg_data.msg.bot.delete_messages(chat_id=msg_data.tgid, message_ids=messages_to_delete)
+    if messages_to_delete is not None and messages_to_delete:
+        await msg_data.msg.bot.delete_messages(chat_id=msg_data.tgid, message_ids=messages_to_delete)
     
     await msg_data.state.update_data({"report:branch": branch, "report:step": step})
     msg_func = get_msg_func(step, branch)
